Move API tests out
